+++ include/linux/mtd/nand.h
@@ -372,7 +372,7 @@ struct nand_chip {
 	int		bbt_erase_shift;
 	int		chip_shift;
 	int		numchips;
-	unsigned long	chipsize;
+	unsigned long long	chipsize;
 	int		pagemask;
 	int		pagebuf;
 	int		badblockpos;
